[video] Clean up types and export the VideoPlayer class (#28623)
