+++ packages/expo-video/src/VideoPlayer.web.tsx
@@ -5,7 +5,7 @@ import type {
   VideoPlayerEvents,
   VideoPlayerStatus,
   VideoSource,
-} from './VideoView.types';
+} from './VideoPlayer.types';
 
 export function useVideoPlayer(
   source: VideoSource,
@@ -27,7 +27,7 @@ export function getSourceUri(source: VideoSource): string | null {
   return source?.uri ?? null;
 }
 
-export class VideoPlayerWeb
+export default class VideoPlayerWeb
   extends globalThis.expo.SharedObject<VideoPlayerEvents>
   implements VideoPlayer
 {
