+++ packages/expo-video/src/index.ts
@@ -1,6 +1,16 @@
-import Video from './NativeVideoModule';
-
 export { VideoView, isPictureInPictureSupported } from './VideoView';
 export { useVideoPlayer } from './VideoPlayer';
-export { Video };
-export * from './VideoView.types';
+
+export { VideoContentFit, VideoViewProps } from './VideoView.types';
+
+export {
+  VideoPlayer,
+  VideoPlayerEvents,
+  VideoPlayerStatus,
+  VideoSource,
+  PlayerError,
+  VolumeEvent,
+  VideoMetadata,
+  DRMType,
+  DRMOptions,
+} from './VideoPlayer.types';
